Restore packages on project build if it never happened before
To prevent VS going crazy on changing configurations we have to manually restore packages.
To optimize this process we skip restore if restore already happened or for builds outside VS (also should work for Rider),
New property DoInsideBuild is defined to be a mark of build from IDE, it can me changed in User.Directory.Build.props
in case some specific ide is needed to be detected.