Add cache of root method call expressions
For every Query.All<T> one .MakeGenericMethod required and the result is only visited.
The same root expressions are re-used many times during domain lifetime
The number of items in the cache is finite and limit is the number of entities and interfaces
in domain model (auxilary types included). In future we may change type of cache from concurrent
dictionary to LRU cache and limit it to have only hot query roots represented