ValueFieldAccessorResolver: Getting value accessors re-arranged
Some types are more frequent to get accessor than others.
For example
- long, int and guid are frequent key values, entity version values so
  one of them will be in every query.
- boolean fields are probably also frequent type amont model fields
- DateTime/DateTimeOffset can be a time stamp for each row.
- Double and decimal values are about money/amout of something/etc, they
  are more important than some other types.

At the same time unsigned integers and Byte/SByte are, to me, used
less frequently.

It's a small optimization but difference between finding accessor on
first and last check is ~2.5x and we always had to make 5-6 checks before
reaching long which is, i believe, the most frequent key value type.